feat: allow deletion of user submissions
